requirements.txt
streamlit
pandas
Pillow
plotly
sqlalchemy
